requirements.txt
fastapi
uvicorn
google-auth-oauthlib
google-auth
google-api-python-client
python-dateutil
redis
google-auth-httplib2